All full testing suite (can be improved further)

+++ Implementation/payfriend_service/service/classes/component/pay/pay.py
@@ -14,11 +14,16 @@ class PaymentComponent():
         self.transactions_context = TransactionsContext()
 
     def process_payment(self, email: str, password: str, value: float, company: str):
-        # Check that user is registered with service
+        # Check that user is registered with the service
         auth = AuthComponent()
         registered = auth.authenticate_user(email, password)
-        if (registered == False):
-            return None, f"Account credentials for {email} are incorrect."
+        if (registered < 0):
+            if registered == -1:
+                return None, f"Provided account credentials are incorrect."
+            elif registered == -2:
+                return None, f"Provided email address is invalid."
+            elif registered == -3:
+                return None, f"Provided password is invalid."
         # Check that payment details are valid
         if (float(value) <= 0):
             return None, f"Payment details not valid (Transaction value)."
@@ -37,6 +42,9 @@ class PaymentComponent():
         return transaction, None
 
     def delete_payment(self, transactionId: str):
+        # Input validation
+        if (transactionId == ""):
+            return False
         status = self.transactions_context.delete_transaction_from_table(transactionId)
         if (status == -1):
             return False
